chore: refactor topics graph data and types into separate files and resolve conflicts
